Refactor and tidy body parsing middleware for WS + Arc v5 VTL emulation
